requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==4.0.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
mysqlclient==2.1.1
PyJWT==2.6.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
